+++ src/buttondevteam/bucket/core/TBMCPlayerAddEvent.java
@@ -1,4 +1,4 @@
-package buttondevteam.core;
+package buttondevteam.bucket.core;
 
 import org.bukkit.event.Event;
 import org.bukkit.event.HandlerList;
